Add URL_KEY to .env.example and fetch author info using BlogService
