+++ .env.example
@@ -1,3 +1,4 @@
 SERVICE_ID=your_service_id
 TEMPLATE_ID=your_template_id
 PUBLIC_KEY=your_public_key
+URL_KEY=your_url_key
