+++ nuxt.config.ts
@@ -16,6 +16,7 @@ export default defineNuxtConfig({
       serviceId: process.env.SERVICE_ID,
       templateId: process.env.TEMPLATE_ID,
       publicKey: process.env.PUBLIC_KEY,  
+      urlKey: process.env.URL_KEY,
     }
   },
   alias: {
